+++ packages/components/components/elvis-tabs/src/react/elvia-tabs.tsx
@@ -109,8 +109,9 @@ const Tabs: FC<TabsProps> = ({ items, value = 0, isInverted, valueOnChange, webc
     ['ewc-tabs__arrow--remove']: isOnRightEnd && isOnLeftEnd,
   });
   const itemsClasses = classNames('ewc-tabs__items', {
-    ['ewc-tabs--hide-fade-right']: isOnRightEnd,
-    ['ewc-tabs--hide-fade-left']: isOnLeftEnd,
+    ['right-arrow-fade']: !isOnRightEnd,
+    ['left-arrow-fade']: !isOnLeftEnd,
+    ['both-arrows-fade']: !isOnLeftEnd && !isOnRightEnd,
     ['ewc-tabs--scrolling']: !isOnLeftEnd || !isOnRightEnd,
   });