+++ packages/components/frameworks/react-ssr/src/App.test.js
@@ -1,8 +1,11 @@
-import { render, screen } from '@testing-library/react';
-import App from './App';
+import '@testing-library/jest-dom';
+import React from 'react';
+import 'jsdom-global/register';
+import { Accordion } from '@elvia/elvis-accordion/react';
+import enzyme from 'enzyme';
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+enzyme.configure({ adapter: new Adapter() });
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+it('should render my component', () => {
+  const wrapper = enzyme.shallow(<Accordion />);
 });