+++ packages/components/components/elvis-chip/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Elvia Chip Changelog
 
+## 1.1.0 (03.11.21)
+
+### New features
+
+- Added type declaration (for react)
+- Bug fix for disabled text
+
+## 1.0.4 (27.10.21)
+
+### New features
+
+- Explicitly set text color in in chip to fix bug in iOS 15.
+
 ## 1.0.3 (03.09.21)
 
 ### New features