+++ packages/components/components/elvis-progress-linear/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Progress-linear Changelog
 
+## 1.1.0 (03.11.21)
+
+### New features
+
+- Added type declaration (for react)
+
 ## 1.0.0
 
 First stable version :tada: